Windows support Whitebox AES 512-16k

+++ src/crypto_ecckey.hpp
@@ -369,7 +369,7 @@ namespace cryptoAL
     public:
 		bool generate_private_public_key(bool verb = false)
 		{
-			long long Nbytes = 1.33 * dom.key_size_bits / 8;
+			long long Nbytes = (long long) 1.33 * dom.key_size_bits / 8;
 			s_k = generate_base64_random_string(Nbytes);
 
 			if (verb) std::cout << "private key:  " << s_k << std::endl;
